% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chea3_query.R
\name{chea3_query}
\alias{chea3_query}
\title{Query ChEA3 API for TF enrichment}
\usage{
chea3_query(genes, query_name = "rChEA3_query")
}
\arguments{
\item{genes}{Character vector of HGNC gene symbols.}

\item{query_name}{Optional query name (default: "rChEA3_query").}
}
\value{
A named list of data frames. Each element corresponds to a ChEA3
collection and contains an enrichment table with transcription factors and
their statistics. The expected names are:
c("Integrated--meanRank", "Integrated--topRank",
"GTEx--Coexpression", "ARCHS4--Coexpression",
"ENCODE--ChIP-seq", "ReMap--ChIP-seq",
"Literature--ChIP-seq", "Enrichr--Queries").
}
\description{
Query ChEA3 API for TF enrichment
}
\examples{
\dontrun{
results <- chea3_query(c("SMAD9","FOXO1","MYC","STAT1","STAT3","SMAD3"))
}
}
